refactor(list): use drill lists for user and official lists
