+++ projects/client/src/lib/sections/lists/components/list-summary/ListPreview.svelte
@@ -1,35 +1,28 @@
 <script lang="ts">
-  import SectionList from "$lib/components/lists/section-list/SectionList.svelte";
   import * as m from "$lib/features/i18n/messages.ts";
+
   import type { MediaListSummary } from "$lib/requests/models/MediaListSummary.ts";
   import type { MediaType } from "$lib/requests/models/MediaType";
   import { useListItems } from "$lib/sections/lists/components/list-summary/_internal/useListItems.ts";
-  import MediaCard from "$lib/sections/lists/components/MediaCard.svelte";
-  import { mediaListHeightResolver } from "$lib/sections/lists/utils/mediaListHeightResolver.ts";
-  import ViewAllButton from "../ViewAllButton.svelte";
+  import DrillableMediaList from "../../drilldown/DrillableMediaList.svelte";
+  import { mediaListHeightResolver } from "../../utils/mediaListHeightResolver";
+  import MediaCard from "../MediaCard.svelte";
   import { getListUrl } from "./_internal/getListUrl";
 
   const { list, type }: { list: MediaListSummary; type?: MediaType } = $props();
-
-  const { items, isLoading } = $derived(useListItems({ list, type }));
-  const isEmptyList = $derived(!$isLoading && $items.length === 0);
+  $inspect(type);
 </script>
 
-<!-- FIXME switch to drillable list when support is added -->
-<SectionList
+<DrillableMediaList
+  {type}
   id={`top-list-${type}-${list.id}`}
-  items={$items}
+  drilldownLabel={m.view_all()}
+  useList={(params) => useListItems({ list, ...params })}
+  urlBuilder={() => getListUrl(list, type)}
   title={list.name}
   --height-list={mediaListHeightResolver(type)}
 >
-  {#snippet actions()}
-    <ViewAllButton
-      href={getListUrl(list, type)}
-      label={m.view_all()}
-      isDisabled={isEmptyList}
-    />
-  {/snippet}
   {#snippet item(media)}
     <MediaCard type={media.entry.type} media={media.entry} />
   {/snippet}
-</SectionList>
+</DrillableMediaList>
