+++ projects/client/src/lib/sections/lists/official/OfficialListPaginatedList.svelte
@@ -1,31 +1,28 @@
 <script lang="ts">
   import Preview from "$lib/components/badge/Preview.svelte";
-  import GridList from "$lib/components/lists/grid-list/GridList.svelte";
   import type { MediaType } from "$lib/requests/models/MediaType";
   import { useMedia, WellKnownMediaQuery } from "$lib/stores/css/useMedia";
   import MediaCard from "../components/MediaCard.svelte";
-  import { useUserList } from "./useUserList";
+  import DrilledMediaList from "../drilldown/DrilledMediaList.svelte";
+  import { useOfficialList } from "./useOfficialList";
 
-  type UserListProps = {
+  type OfficialListProps = {
     title: string;
-    userId: string;
     listId: string;
     type?: MediaType;
   };
 
-  const { title, userId, listId, type }: UserListProps = $props();
+  const { title, listId, type }: OfficialListProps = $props();
 
-  const { list } = $derived(useUserList({ userId, listId, type }));
   const isMobile = useMedia(WellKnownMediaQuery.mobile);
   const style = $derived($isMobile ? "summary" : "cover");
 </script>
 
-<!-- TODO use drilled media list & fetch rest on scroll -->
-<GridList
-  id={`userlist-list-${userId}-${listId}`}
+<DrilledMediaList
+  id={`official-list-${listId}`}
   {title}
-  items={$list}
-  --width-item="var(--width-poster-card)"
+  {type}
+  useList={(params) => useOfficialList({ listId, ...params })}
 >
   {#snippet item(media)}
     <MediaCard type={media.type} {media} {style} />
@@ -34,4 +31,4 @@
   {#snippet badge()}
     <Preview />
   {/snippet}
-</GridList>
+</DrilledMediaList>
