+++ projects/client/src/routes/lists/official/[id]/+page.svelte
@@ -2,7 +2,7 @@
   import { page } from "$app/state";
   import TraktPage from "$lib/sections/layout/TraktPage.svelte";
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
-  import OfficialList from "$lib/sections/lists/official/OfficialList.svelte";
+  import OfficialListPaginatedList from "$lib/sections/lists/official/OfficialListPaginatedList.svelte";
   import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
   import { mapToMediaType } from "../../../users/[user]/lists/[list]/_internal/mapToMediaType";
   import { useListSummary } from "./useListSummary";
@@ -21,5 +21,5 @@
 <TraktPage audience="all" image={DEFAULT_SHARE_COVER} title={listName}>
   <TraktPageCoverSetter />
 
-  <OfficialList title={listName} listId={page.params.id} {type} />
+  <OfficialListPaginatedList title={listName} listId={page.params.id} {type} />
 </TraktPage>
