+++ projects/client/src/routes/users/[user]/lists/[list]/+page.svelte
@@ -2,7 +2,7 @@
   import { page } from "$app/state";
   import TraktPage from "$lib/sections/layout/TraktPage.svelte";
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
-  import UserList from "$lib/sections/lists/user/UserList.svelte";
+  import UserListPaginatedList from "$lib/sections/lists/user/UserListPaginatedList.svelte";
   import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
   import { mapToMediaType } from "./_internal/mapToMediaType";
   import { userListSummary } from "./userListSummary.ts";
@@ -21,7 +21,7 @@
 <TraktPage audience="all" image={DEFAULT_SHARE_COVER} title={listName}>
   <TraktPageCoverSetter />
 
-  <UserList
+  <UserListPaginatedList
     title={listName}
     userId={page.params.user}
     listId={page.params.list}
